docs: move the appwrite and keyboard shortcuts pages under command center articles, add active keyboard shortcuts table, move paragraphs around a bit in the appwriter article
