+++ src/redirects.json
@@ -661,7 +661,7 @@
     },
     {
         "link": "/keyboard",
-        "redirect": "/docs/tooling/assistant/appwriter"
+        "redirect": "/docs/tooling/command-center/the-appwriter"
     },
     {
         "link": "/careers",
