+++ src/routes/docs/advanced/platform/+layout.svelte
@@ -14,10 +14,6 @@
                 {
                     label: 'Overview',
                     href: '/docs/advanced/platform'
-                },
-                {
-                    label: 'Shortcuts',
-                    href: '/docs/advanced/platform/shortcuts'
                 }
             ]
         },
